chore: move form-qualifiedAttributions component
