+++ src/components/form-qualifiedAttributions/configured-form-qualified-attributions.ts
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import { reduxForm } from 'redux-form';
 
 import { FormQualifiedAttributions } from './form-qualified-attributions.component';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'qualifiedAttributions',
